requirements.txt
GitPython
python-dateutil
mermaid
Pillow